requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
Pillow>=9.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.20.0
gradio>=6.2.0
huggingface-hub>=0.20.0